requirements.txt
Flask==3.0.0
flask-cors==4.0.0
rembg==2.0.50
Pillow==10.1.0
python-dotenv==1.0.0
gunicorn==21.2.0
